fix #76 - Fix: type 분리
